separate files for bridgezone resource
