+++ service/controller/v24/resource/bridgezone/resource.go
@@ -9,13 +9,11 @@ import (
 	"github.com/aws/aws-sdk-go/service/route53"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"golang.org/x/sync/errgroup"
 	"k8s.io/client-go/kubernetes"
 
 	clientaws "github.com/giantswarm/aws-operator/client/aws"
 	"github.com/giantswarm/aws-operator/service/controller/v24/controllercontext"
 	"github.com/giantswarm/aws-operator/service/controller/v24/credential"
-	"github.com/giantswarm/aws-operator/service/controller/v24/key"
 )
 
 const (
@@ -146,228 +144,6 @@ func New(config Config) (*Resource, error) {
 
 func (r *Resource) Name() string {
 	return name
-}
-
-func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	customObject, err := key.ToCustomObject(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if !r.route53Enabled {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "route53 disabled")
-		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource reconciliation for custom object")
-		return nil
-	}
-
-	baseDomain := key.BaseDomain(customObject)
-	intermediateZone := "k8s." + baseDomain
-	finalZone := key.ClusterID(customObject) + ".k8s." + baseDomain
-
-	guest, defaultGuest, err := r.route53Clients(ctx)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	g := &errgroup.Group{}
-
-	var intermediateZoneID string
-	g.Go(func() error {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "getting intermediate zone ID")
-
-		id, err := r.findHostedZoneID(ctx, defaultGuest, intermediateZone)
-		if IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "intermediate zone not found")
-
-			return microerror.Mask(err)
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-		intermediateZoneID = id
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "got intermediate zone ID")
-
-		return nil
-	})
-
-	var finalZoneID string
-	g.Go(func() error {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "getting final zone ID")
-
-		id, err := r.findHostedZoneID(ctx, guest, finalZone)
-		if IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "final zone not found")
-
-			return microerror.Mask(err)
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-		finalZoneID = id
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "got final zone ID")
-
-		return nil
-	})
-
-	err = g.Wait()
-	if IsNotFound(err) {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
-
-		return nil
-	} else if err != nil {
-		return microerror.Mask(err)
-	}
-
-	var finalZoneRecords []*route53.ResourceRecord
-	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", "getting final zone name servers")
-
-		nameServers, _, err := r.getNameServersAndTTL(ctx, guest, finalZoneID, finalZone)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		for _, ns := range nameServers {
-			copy := ns
-			v := &route53.ResourceRecord{
-				Value: &copy,
-			}
-			finalZoneRecords = append(finalZoneRecords, v)
-		}
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "got final zone name servers")
-	}
-
-	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", "ensuring final zone delegation from intermediate zone")
-
-		upsert := route53.ChangeActionUpsert
-		ns := route53.RRTypeNs
-		ttl := int64(900)
-
-		in := &route53.ChangeResourceRecordSetsInput{
-			ChangeBatch: &route53.ChangeBatch{
-				Changes: []*route53.Change{
-					{
-						Action: &upsert,
-						ResourceRecordSet: &route53.ResourceRecordSet{
-							Name:            &finalZone,
-							Type:            &ns,
-							TTL:             &ttl,
-							ResourceRecords: finalZoneRecords,
-						},
-					},
-				},
-			},
-			HostedZoneId: &intermediateZoneID,
-		}
-		_, err := defaultGuest.ChangeResourceRecordSetsWithContext(ctx, in)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "ensured final zone delegation from intermediate zone")
-	}
-
-	return nil
-}
-
-func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	customObject, err := key.ToCustomObject(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if !r.route53Enabled {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "route53 disabled")
-		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource reconciliation for custom object")
-		return nil
-	}
-
-	baseDomain := key.BaseDomain(customObject)
-	intermediateZone := "k8s." + baseDomain
-	finalZone := key.ClusterID(customObject) + ".k8s." + baseDomain
-
-	_, defaultGuest, err := r.route53Clients(ctx)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	var intermediateZoneID string
-	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", "getting intermediate zone ID")
-
-		intermediateZoneID, err = r.findHostedZoneID(ctx, defaultGuest, intermediateZone)
-		if IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "intermediate zone not found")
-			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource reconciliation for custom object")
-			return nil
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "got intermediate zone ID")
-	}
-
-	var finalZoneTTL int64
-	var finalZoneRecords []*route53.ResourceRecord
-	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", "getting final zone delegation name servers and TTL from intermediate zone")
-
-		nameServers, ttl, err := r.getNameServersAndTTL(ctx, defaultGuest, intermediateZoneID, finalZone)
-		if IsNotFound(err) {
-			// Delegation may be already deleted. It must be handled.
-			r.logger.LogCtx(ctx, "level", "debug", "message", "final zone delegation not found in intermediate zone")
-			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource reconciliation for custom object")
-			return nil
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-
-		finalZoneTTL = ttl
-
-		for _, ns := range nameServers {
-			copy := ns
-			v := &route53.ResourceRecord{
-				Value: &copy,
-			}
-			finalZoneRecords = append(finalZoneRecords, v)
-		}
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "got final zone delegation name servers and TTL from intermediate zone")
-	}
-
-	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", "ensuring deletion of final zone delegation from intermediate zone")
-
-		delete := route53.ChangeActionDelete
-		ns := route53.RRTypeNs
-
-		in := &route53.ChangeResourceRecordSetsInput{
-			ChangeBatch: &route53.ChangeBatch{
-				Changes: []*route53.Change{
-					{
-						Action: &delete,
-						ResourceRecordSet: &route53.ResourceRecordSet{
-							Name:            &finalZone,
-							Type:            &ns,
-							TTL:             &finalZoneTTL,
-							ResourceRecords: finalZoneRecords,
-						},
-					},
-				},
-			},
-			HostedZoneId: &intermediateZoneID,
-		}
-		_, err := defaultGuest.ChangeResourceRecordSetsWithContext(ctx, in)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		r.logger.LogCtx(ctx, "level", "debug", "message", "ensured deletion of final zone delegation from intermediate zone")
-	}
-
-	return nil
 }
 
 // findHostedZoneID fetches Route53 hosted zone IDs based on a given name. The
